switch to custom fork of bitcoin-rpc-client

+++ paciofs-server/src/main/java/de/zib/paciofs/multichain/rpc/MultiChainRpcClient.java
@@ -31,7 +31,7 @@ public interface MultiChainRpcClient extends BitcoindRpcClient {
 
     String protocol();
 
-    short port();
+    int port();
 
     long setupBlocks();
